#4380 - Refactor (calculatePreviewTop.ts): Rename the file
