+++ packages/ketcher-macromolecules/src/helpers/calculatePreviewTop.ts
